seperate move request tests, work around for consistent item with single page request
